refactor: separate panel management into dedicated module
